Add Indexer for Mobile App
This is a very specific endpoint that is only useful for Cordova Applications that are using the `cordova-plugin-indexappcontent` plugin.  Which is why this endpoint is not documented in the Apiary Docs, as it is not intended for public use, unless ... you know, you are building a Cordova App and want our indexed content for your app ... then, go for it.